// src/client/__generated__/types.ts
export interface CreateCustomerMutation_createCustomer {
  __typename: "Customer";
  id: string;
  name: string | null;
  identificationNumber: string | null;
  personName: string | null;
  email: string | null;
  phone: string | null;
}

export interface CreateCustomerMutation {
  createCustomer: CreateCustomerMutation_createCustomer;
}

export interface CreateCustomerMutationVariables {
  input: CreateCustomerInput;
}
// ...
export interface AddressInput {
  city?: string | null;
  isPrimary?: boolean | null;
  number?: string | null;
  postNumber?: string | null;
  state?: string | null;
  street?: string | null;
}

export interface CreateCustomerInput {
  addresses?: AddressInput[] | null;
  allowedBankPayments?: boolean | null;
  email?: string | null;
  identificationNumber?: string | null;
  name?: string | null;
  note?: string | null;
  personName?: string | null;
  phone?: string | null;
  taxIdentificationNumber?: string | null;
}